refactor: project-detail 컴포넌트를 shared로 이동
- features/project-detail/* → shared/components/project/*로 파일 이동
- 여러 feature에서 공통으로 사용하는 컴포넌트이므로 shared 위치가 적합
- FSD 아키텍처 원칙 준수: features는 독립적 기능, shared는 공통 재사용

변경된 import 경로:
- app/(dashboard)/project/[id]/layout.tsx
- features/wbs/components/WBSTableView.tsx
- shared/components/project/components/ProjectView.tsx (self-reference)

파일 구조:
- shared/components/project/components/ (ProjectHeader, ProjectLayoutClient, ViewNavigator 등)
- shared/components/project/skeletons/ (ProjectViewSkeleton)
