+++ shared/components/project/components/TaskDetailPanel.tsx
