+++ shared/components/project/components/ViewNavigator.tsx
